requirements.txt
google-generativeai
chromadb
sentence-transformers
dotenv
numpy
Django
markdown2
PyPDF2
python-docx
odfpy